migrate closestPointToGeometry
